core: Add Kamakiri2 implementation
Kamakiri2 is a BootROM exploit based on the original Kamakiri by xyzz and chaosmaster,
that abuses the linecode exploit found by Chimera.

This implementation is a middle ground between mtkclient's generic payload and MTK-bypass flow.

Instead of having a lot of different payloads for each SoC, Penumbra's implementation uses
a new, smaller, written from scratch payload which works on every SoC thanks to dynamic search.

Credits and thanks goes to the following people:
* @R0rt1z2 for the idea of handling SoC specific values through a custom header
* @xyzz and @chaosmaster (k4y0z) for the original Kamakiri exploit
* the Chimera team for discovering the linecode exploit
* @bkerler for the original generic patcher implementation in mtkclient