core: Allow flashing and reading other sections
The API was hardcoding (as a consequence of leftover code during testing) the section provided to write_flash and read_flash.
Because of this, it was not possible to use these functions to write and read partitions like `preloader` and `preloader_backup`, and instead would only cause reading and flashing PGPT, which in some cases could lead to a soft brick.