core: Add small RSA encryption utility
All available crates for Rust don't allow RSA OAEP encryption with the PrivateKey.

While not implementing this is a standard for RSA implementations, MediaTek uses the PrivateKey to encrypt SLA challenges, that will then later decrypted by the DA.

While this implementation is not secure for actual encryption, it's good enough for our specific use case due to its simplicity and minimal footprint.